refactor: nickname 관련 UseCase 패키지 위치 변경
- `com.smtm.pickle.domain.usecase.nickname` 패키지를 `com.smtm.pickle.domain.usecase.user`로 이동
- `GetNicknameUseCase`, `SaveNicknameUseCase`, `ObserveNicknameUseCase` 경로 수정 및 관련 import문 업데이트
