+++ presentation/src/main/java/com/smtm/pickle/presentation/home/HomeViewModel.kt
@@ -6,7 +6,7 @@ import com.smtm.pickle.domain.model.ledger.summarize
 import com.smtm.pickle.domain.usecase.ledger.EnsureLedgersSyncedUseCase
 import com.smtm.pickle.domain.usecase.ledger.ObserveLedgersByDayUseCase
 import com.smtm.pickle.domain.usecase.ledger.ObserveLedgersByMonthUseCase
-import com.smtm.pickle.domain.usecase.nickname.ObserveNicknameUseCase
+import com.smtm.pickle.domain.usecase.user.ObserveNicknameUseCase
 import com.smtm.pickle.presentation.common.model.ledger.LedgerUiModel
 import com.smtm.pickle.presentation.common.model.ledger.toUiModel
 import com.smtm.pickle.presentation.home.model.LedgerCalendarDay
@@ -85,7 +85,7 @@ class HomeViewModel @Inject constructor(
             .flatMapLatest { yearMonth ->
                 ensureLedgersSynced(yearMonth)
                 observeLedgersByMonthUseCase(yearMonth)
-                    .map { ledgers -> yearMonth to ledgers}
+                    .map { ledgers -> yearMonth to ledgers }
                     .catch { e ->
                         Timber.e(e, "observeLedgersByMonthUseCase() failed")
                         _effect.emit(HomeEffect.ShowSnackBar("데이터를 불러오는데 실패했습니다."))
