+++ presentation/src/main/java/com/smtm/pickle/presentation/login/nickname/NicknameViewModel.kt
@@ -2,7 +2,7 @@ package com.smtm.pickle.presentation.login.nickname
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.smtm.pickle.domain.usecase.nickname.SaveNicknameUseCase
+import com.smtm.pickle.domain.usecase.user.SaveNicknameUseCase
 import com.smtm.pickle.presentation.common.constant.NicknameValidation.MAX_NICKNAME_LENGTH
 import com.smtm.pickle.presentation.common.utils.NicknameUtils
 import dagger.hilt.android.lifecycle.HiltViewModel
