+++ presentation/src/main/java/com/smtm/pickle/presentation/mypage/MyPageViewModel.kt
@@ -7,7 +7,7 @@ import com.smtm.pickle.domain.model.ledger.LedgerType
 import com.smtm.pickle.domain.model.ledger.summarize
 import com.smtm.pickle.domain.usecase.ledger.EnsureLedgersSyncedUseCase
 import com.smtm.pickle.domain.usecase.ledger.ObserveLedgersByMonthUseCase
-import com.smtm.pickle.domain.usecase.nickname.ObserveNicknameUseCase
+import com.smtm.pickle.domain.usecase.user.ObserveNicknameUseCase
 import com.smtm.pickle.presentation.common.model.ledger.toUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
@@ -31,7 +31,7 @@ class MyPageViewModel @Inject constructor(
     private val _uiState = MutableStateFlow(MyPageUiState())
     val uiState: StateFlow<MyPageUiState> = _uiState.asStateFlow()
 
-    private val _effect : Channel<MyPageEffect> = Channel(Channel.BUFFERED)
+    private val _effect: Channel<MyPageEffect> = Channel(Channel.BUFFERED)
     val effect = _effect.receiveAsFlow()
 
     init {
