+++ presentation/src/main/java/com/smtm/pickle/presentation/mypage/profile/ProfileViewModel.kt
@@ -2,7 +2,7 @@ package com.smtm.pickle.presentation.mypage.profile
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.smtm.pickle.domain.usecase.nickname.ObserveNicknameUseCase
+import com.smtm.pickle.domain.usecase.user.ObserveNicknameUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
